#7 Updating dependencies and taking server-test-s…
…upport library into use.
