+++ server/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ty-ras/server-koa",
-  "version": "0.9.0",
+  "version": "0.10.0",
   "author": {
     "name": "Stanislav Muhametsin",
     "email": "346799+stazz@users.noreply.github.com",
@@ -13,33 +13,37 @@
   },
   "files": [
     "./src",
-    "./dist",
+    "./dist-mjs",
+    "./dist-cjs",
     "README.md",
     "LICENSE.txt"
   ],
   "type": "module",
-  "main": "./dist/index.js",
-  "module": "./dist/index.mjs",
+  "main": "./dist-cjs/index.js",
+  "module": "./dist-mjs/index.mjs",
   "types": "./src/index.ts",
   "exports": {
     ".": {
       "types": "./src/index.ts",
-      "require": "./dist/index.js",
-      "import": "./dist/index.mjs"
+      "import": "./dist-mjs/index.mjs",
+      "require": "./dist-cjs/index.js"
     }
   },
   "dependencies": {
-    "@ty-ras/server": "0.9.4",
-    "@ty-ras/endpoint-prefix": "0.9.0"
+    "@ty-ras/server": "0.10.1",
+    "@ty-ras/endpoint-prefix": "0.10.1"
   },
   "peerDependencies": {
-    "koa": "2.13.4"
+    "koa": "^2.13.4"
   },
   "devDependencies": {
+    "@babel/core": "7.19.3",
+    "@babel/eslint-parser": "7.19.1",
     "@types/koa": "2.13.5",
     "@types/node": "18.7.18",
     "@typescript-eslint/eslint-plugin": "5.38.0",
     "@typescript-eslint/parser": "5.38.0",
+    "@ty-ras/server-test-support": "0.10.5",
     "ava": "4.3.3",
     "@ava/get-port": "2.0.0",
     "c8": "7.12.0",
@@ -55,12 +59,12 @@
   },
   "scripts": {
     "build:run": "yarn run lint && yarn run tsc",
-    "build:ci": "yarn run clear-build-artifacts && yarn run tsc --project tsconfig.build.json && yarn run rename-js-to-mjs && yarn run tsc --project tsconfig.build.json --module CommonJS",
-    "clear-build-artifacts": "rm -rf dist build",
-    "format-output-files": "eslint --no-eslintrc --config '.eslintrc.output.cjs' --fix 'dist/*js'",
+    "build:ci": "yarn run clear-build-artifacts && yarn run tsc --outDir ./dist-mjs && yarn run rename-js-to-mjs && yarn run tsc --module CommonJS --outDir ./dist-cjs && yarn run format-output-files",
+    "clear-build-artifacts": "rm -rf dist dist-cjs dist-mjs build",
+    "format-output-files": "eslint --no-eslintrc --config '.eslintrc.output.cjs' --fix 'dist-cjs/*js' 'dist-mjs/*js'",
     "lint": "eslint ./src --ext .ts,.tsx",
-    "rename-js-to-mjs": "find dist -name '*.js' -type f -exec sh -c 'for f in \"$@\"; do mv -- \"$f\" \"${f%.js}.mjs\"; done' -- {} +",
-    "tsc": "tsc",
+    "rename-js-to-mjs": "find dist-mjs -name '*.js' -type f -exec sh -c 'for f in \"$@\"; do mv -- \"$f\" \"${f%.js}.mjs\"; done' -- {} +",
+    "tsc": "tsc --project tsconfig.build.json",
     "test:coverage": "c8 ava",
     "test:run": "ava"
   }
