+++ tsconfig.build.json
@@ -9,4 +9,8 @@
     // Also, this way, we will not compile any of the tests files, thus avoiding need to move them around in CI pipeline.
     "src/**/__test__/*"
   ],
+  "compilerOptions": {
+    // We don't want dangling // eslint-disable-xyz comments, as that will cause errors during formatting output .[m]js files.
+    "removeComments": true
+  },
 }
